fixed data type

+++ messenger/src/main/java/com/jtjmpm/messages/PlayerDTO.java
@@ -2,6 +2,6 @@ package com.jtjmpm.messages;
 
 public record PlayerDTO(
         String id,
-        int hp,
+        double hp,
         boolean ready
 ) {}
